[AIM] Temporary fix for AIM searchbox colors in GTK Theme
On Linux, when GTK theme is enabled, the composebox element color
provider marks the page as dark mode, which mismatches the values Lens
gets from the ThemeService. Long term, we should prevent this mismatch,
so we can rely on the colors provided by the composebox element.
However, for the short term, and as a cherry pick fix, I am hard coding
in the values in the Lens file to prevent the composebox using dark
color tokens while Lens is in light mode (or vice versa).

Bug: b:449214753
Change-Id: I5c3ffd671e74b5c75a9a3d2bfecc9aaf4ae2c7bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7014001
Reviewed-by: Juan Mojica <juanmojica@google.com>
Commit-Queue: Duncan Mercer <mercerd@google.com>
Cr-Commit-Position: refs/heads/main@{#1525941}