Add guard against invalid transform node ids
The code that changes tile tree priorities for checkerboarding is
currently accessing the transform tree's nodes assuming the currently
scrolling node has a valid `transform_id`. This assumption does not
always hold and in some scenarios the code was trying to access an
invalid node in the tree, causing an exception to be thrown.

This CL skips the code that verifies if there is checkerboarding if the
scroll node doesn't have a valid transform_id, since this means that the
scroll is starting, and therefore checkerboarding is not possible yet.
This CL also adds to the CHECKs in the functions that access transform
tree nodes to verify that they have a valid id.

This scenario occurred under very particular situations:
1. The scroll needs to be a gesture scroll, and the gesture has to be
initiated via touch.
2. Overlay scrollbars need to be enabled, Fluent or not.
3. This only happens at the start of a scroll; once a scroll is ongoing
the scroll_node will have a proper transform_id set.

Bug: 443348409
Change-Id: I11af872ca7026c931ce3088732505cf1d1812482
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7006657
Reviewed-by: Olga Gerchikov <gerchiko@microsoft.com>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Gaston Rodriguez <gastonr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1525944}