[AIM] Fix submit icon being too big
The icon was accidentally being overrode by the ::part(action-icon). It
was being set to 24px when it should've been 16px, which is defined in
the composebox.html.ts file.

This has crept into Stable, and so I am keeping this CL simple in hopes
of CPing back.

Bug: b:449814941
Change-Id: I8252ace168876fbc71019e5db01aea8ae2d8cca1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7014868
Reviewed-by: Bryan Nguyen <nguyenbryan@google.com>
Commit-Queue: Duncan Mercer <mercerd@google.com>
Cr-Commit-Position: refs/heads/main@{#1525967}