[High Contrast] [win] Initialize high contrast(forced color) state during the OsSettingsProviderWin construction.
Key change: Sets the forced_colors_active_ member variable based on
whether the HCF_HIGHCONTRASTON flag is set in the result. This change
ensures that the high contrast state is properly initialized when the
OsSettingsProviderWin object is created, which is important in the
Windows UI theming system.

Bug:448235813

Change-Id: I7ff338ec014d9c8954148c97c8e283948aeb8ff4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6997603
Reviewed-by: Tom Lukaszewicz <tluk@chromium.org>
Commit-Queue: Sejal Anand <sejalanand@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1526034}