Implement SyncServiceObserver::OnSyncShutdown in c/b/ash/
Specifically:

SyncAppsyncOptinClient: The implementation can simply be NOTREACHED(),
since this class must be destroyed before SyncService shutdown (per ctor
precondition comment, and I've also verified manually that this is
indeed the case in practice).

SyncConsentScreen: OnSyncShutdown() is *probably* not reached in
practice, but it's hard to be sure, and it's safe & easy to reset the
SyncService observation.

The eventual goal is to make SyncServiceObserver::OnSyncShutdown() pure
virtual, so that all implementations need to consider whether they need
to do something there (i.e. unregister themselves, drop and SyncService
pointers).

Bug: 40882734
Change-Id: Ia3b712bcdb3eb77fc2670c2c3b1c0231fff5aa11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7011294
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1526142}