Fix the flaky testClearAbusiveNotificationPermissionsReviewList
The stack trace ends with a failure in expectation that an element should be visible while it isn't. This is most likely just a race condition in the test setup - we're testing the element before it's made visible.

We'll handle this case by using onViewWaiting(), which should optimistically fix the flakiness.

Note onViewWaiting() specifically waits for an element to be made visible, thus we're keeping the "not visible" arm of this test synchronous.

Bug: 390581526
Change-Id: I3aa557d0d6eaea4a54f95db7c08df1fa7eb1f36b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7011689
Reviewed-by: Zaina Al-Mashni <zalmashni@google.com>
Commit-Queue: Martin Šrámek <msramek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1526173}