Fix the account display in the Safety Hub to RTL
SafetyHubNoCompromisedPasswordsModuleHelper.getSummary() concatenates an email (always LTR) with a translated text (which could be LTR or RTL). This causes the email to be left-aligned.

We wrap the account in BidiFormatter.unicodeWrap(), which adds Unicode control character `LRE` (left-to-right embedding) that should correctly display the LTR text in a RTL string.

Bug: 433646162
Change-Id: I3f956dcd4f3d7dbaada7c7b5f186d703fbd5c9fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7013588
Reviewed-by: Zaina Al-Mashni <zalmashni@google.com>
Commit-Queue: Martin Šrámek <msramek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1526181}