[ios] Move kCreateTabHelperOnlyForRealizedWebStates to //ios/web
With the feature defined in //ios/chrome/browser, it is not possible
to detect violation of the contract (i.e. creating a tab helper for
an unrealized WebState) when the tab helper is created but only when
the WebState is realized.

This makes it difficult to debug those issues, especially if there is
no known reproduction steps. Moving the feature into //ios/web will
allow to CHECK(...) in WebStateUserData<T>::CreateForWebState(...)
that the WebState is not realized, thus allowing to detect the issue
at the point it happens.

The current CL only moves the feature definition. Enforcement will
be done in a following CL.

Bug: 409299519, 449265570
Change-Id: I836f1ac10c3a7fc65ba948f44f67435b9f12b917
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7006159
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1526205}