[ios][web] Forbids creation of tab helper for unrealized WebStates
Forbids creation of tab helper for unrealized WebStates when the
feature kCreateTabHelperOnlyForRealizedWebStates is enabled. This
allows detecting the violation when they occurs, simplifying the
debugging process (previously the error would only be reported at
the time of realization, which made it difficult to investigate
from a crash happening in production).

Bug: 409299519, 449265570
Change-Id: Ic7162a2ca41764929ea992d117793b6f75f87e8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7002796
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1526209}