Do not add confirmation bubbles to the bubble manager queue
The confirmation bubbles are either time bounded or their state gets
wiped out when the bubble is closed for some reason. They are
implemented with the implicit assumption to be shown once, but with the
BubbleManager, the bubbles can be shown multiple times. This change
ensures that the confirmation bubbles are shown at-most once.

Bug: 432429605
Change-Id: Iecf586b3494269f64e8db21cc782de44023f158b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7004668
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Vidhan Jain <vidhanj@google.com>
Cr-Commit-Position: refs/heads/main@{#1526228}