[Safe Browsing] Fix WebUI AddCookie error
Within the AddCookie() function in the safe_browsing.ts file, an assert
statement intends to check that the variable 'secondResult' is non-null.
However, because the variable can be a number, namely 0, the assert
statement is unintentionally firing.

In this CL, the assert statement is modified to strictly check whether
the variable contains a value.

Bug: 447473363

Change-Id: I13fa051643137cb8e4ae0f50d549ff4e31a5a87b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7011169
Reviewed-by: thefrog <thefrog@chromium.org>
Commit-Queue: Daniel White <danieltwhite@google.com>
Auto-Submit: Daniel White <danieltwhite@google.com>
Cr-Commit-Position: refs/heads/main@{#1526299}