Enable setting ECT(1) on UDP packets in P2PSocketUdp.
While I'm here, I'm also changing the calls to
`socket_->SetDiffServCodePoint()` to use `socket_->SetTos()` instead.
The former is just a wrapper for the latter.

Before this change, there was logic that skipped attempting to set the
DiffServ codepoint if there was a non-transient error *and* setting the
DiffServ codepoint succeeded in the past [0]. The logic as implemented
was in direct contradiction of the comments, which claimed that an
attempt would be skipped if there was a non-transient error and previous
calls had *not* succeeded.

I've replaced this logic with an exponential backoff on non-transient
failure. UMA logging has also been added.

[0]
http://source.chromium.org/chromium/chromium/src/+/1528253a8624e9ecc3e0dd3f5074721772d8be6c:services/network/p2p/socket_udp.cc;l=398

Change-Id: I553098ec596d8864c9f5a08fb62a53fb83e9d894
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6968629
Reviewed-by: Harald Alvestrand <hta@chromium.org>
Reviewed-by: Johannes Kron <kron@chromium.org>
Commit-Queue: Frederik Hermans <frederikh@google.com>
Cr-Commit-Position: refs/heads/main@{#1526313}