Move an early return in ASG from the Generate to the Fetch phase
As mentioned in the interface documentation [1], all filtering of
suggestion data should happen in SG::FetchSuggestionData() since it is
important for all SGs to know that if a certain SG returned data then it
is ready to show suggestions (This will be crucial during the
implementation of the prioritization logic).

AutocompleteSuggestionGenerator contained a filtering step in the
Generation phase, therefore this CL moves it back to the Fetch phase.

[1] https://source.chromium.org/chromium/chromium/src/+/main:components/autofill/core/browser/suggestions/suggestion_generator.h;l=100-104;drc=387d21260098af6cca5369bdb56632330ae2e9c3

Bug: 409962888
Change-Id: I63e2fdea9833d9c8a0c02795798150bc16063ecd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7015137
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Reviewed-by: Karol Sygiet <sygiet@google.com>
Cr-Commit-Position: refs/heads/main@{#1526319}