Add net::UrlSearchParamsView::SerializeAsUtf8()
Add a method to serialize a UrlSearchParamsView as a UTF-8 string. The
serialization preserves equality. This is useful when using it as a hash
key.

It would be possible to write a hash function for UrlSearchParamsView
as-is, but it would be inefficient in practice.

BUG=403211337

Change-Id: I750d61810ce7f48e713fad32cd336df5a6ea1369
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7006228
Commit-Queue: Adam Rice <ricea@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1526327}