[iOS] Add egtests for home background customization
This CL adds egtests for background customization. To enable that, it
splits out a few constant files from their main targets, as there are
many things that eg tests can't depend on. It also introduces many new
accessibility identifiers for the tests to use.

Finally, one of the features uses PHPickerViewController, which can't be
controlled with EarlGrey, so this test uses XCUITest to control the
picker.

Bug: 447151991
Change-Id: I2795228e593d6ec827f4f8409408e52e3b36df5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7008710
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Cheick Cisse <cheickcisse@google.com>
Reviewed-by: Adam Arcaro <adamta@google.com>
Cr-Commit-Position: refs/heads/main@{#1526359}