Update Corp Message to send "1000000 squirrels" message to the client
This feature is to test messages with large payloads, here we
generate a payload with 1,000,000 squirrels and send it to the
client. This should be an upper bound for any message we could
legitimately send so it seems like a reasonable ceiling.

Bug: 359620500
Change-Id: I503ede346f36f9637638e45ffcf1563a86344e14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7017749
Auto-Submit: Joe Downing <joedow@chromium.org>
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Commit-Queue: Jamie Walch <jamiewalch@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1526361}