[ios] Only call -blurComplete & -unblurComplete at the end of animation
The documentation for -animateWithDuration:animations:completion: does
not specify whether it is possible for completion block to be called
with `NO` while the animation is in progress and then with `YES` when
the animation is complete.

In case this is possible this would lead to a crash because the method
-unblurComplete would be called twice, and resulting in a call to the
method -deactivateOverlay: with nil as a parameter.

Fix the completion blocks passed to UIKit to only call -blurComplete &
-unblurComplete if the `complete` parameter is `YES`.

Bug: 444179164
Change-Id: I1fd2213e2b71fa8032c342f3b45e891ae8863b61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7013214
Commit-Queue: Federica Germinario <fedegermi@google.com>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Federica Germinario <fedegermi@google.com>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1526362}