[BrowserView] Clean up logic for adding the BookmarkBarView
Always add the bookmark bar to the top_container view as that is the
only view it should be added to and removed from. This allows us to
simplify the logic in BrowserView::MaybeShowBookmarkBar.

For added context, while attempting to remove the detach logic for the
bookmark bar it was made apparent that this code does not refer to the
old detached bookmark bar feature. Instead it refers to adding the
bookmark bar view from the view hierarchy as a form of lazy loading.

It is possible to remove this logic entirely under the condition that
BrowserViewLayout no longer manages the visibility of the bookmark
button. It's not immediately clear if that is the right direction to go
in at this point in time.

Change-Id: I3fdd154cf5fd7a76427fce490da0f022c306317a
Bug: 449006672
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7014262
Reviewed-by: Christopher Grant <cjgrant@chromium.org>
Commit-Queue: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1526392}