Change underline updater to rely on sharing manager to handle IsShowing.
The previous implementation of underline decoration checked if the Glic
window was showing to control underlines. In the multi-instance
implementation we want this concern to be wholly-owned by the umbrella
sharing manager (via the pinned tabs api). For now, it checks the
feature flag. Once everything is multi-instance, we can remove the
checks and the old IsGlicWindowShowing method entirely.

Bug: 444463509
Change-Id: I65022f09d5f4d6a0aa76673ee00f5bfc32c26d6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7014631
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Marke Hallowell <wry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1526395}