[Predictable App Update] Add throttle for silent icon updates
As per the new predictable app update algorithm, app updates where the
icons have a less than 10% diff are allowed to update silently. As per
the PRD, this CL introduces a throttle that limits that to once per day
instead of multiple ones to prevent a recurring spoofing attack that can
silently update icons over the course of a single day until the attack
succeeds. In case of successive icon updates being triggered within 24
hours, the update gets stored in the pending information for the app.

This is done by caching the last time a silent icon was triggered in the
ManifestUpdateManager and passed in to the manifest silent update
command as an input. Metrics will be updated in a follow up CL.

There could be a potential flaw here where the pending icons get stored
in a pending update info for the web app, but aren't cleared if the icon
updates are applied silently after 24 hours if the user visits the site.
This will also be handled in a follow up CL.

Bug: 448891020
Change-Id: If0395d24e3baf4d5a1a1fc695b1d07d32715d187
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7007487
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: May Siem <msiem@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1526400}