Remove the kLinkedServicesSetting disabled code on Desktop and Android
Since the feature has been launched.

Apart from removing many references to the feature, flag, and tests, we're also adapting the ManageSyncSettings code from the previous approach (where the old strings were set in the XML file and the new strings programmatically overwritten in the Java code) to one where the strings are simply set in the XML file.

We're not removing the equivalent codepaths on iOS, where it hasn't launched yet.

Bug: 343132001
Change-Id: I8b92689b67db9485e87dd99657a32f1601277553
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7006398
Reviewed-by: Rainhard Findling <rainhard@chromium.org>
Commit-Queue: Martin Šrámek <msramek@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1526439}