Keep MIDIAccessInitializer alive during promise resolution
There is a possibility where the MIDIAccessInitializesr can be garbage
collected between StartSession() and DidStartSession(), which would
result in the promise not being properly detached. Keep it alive
instead.

Also add a disconnect handler to properly allow it to be collected if
there is an error.

The reproduction case attached to the bug does not execute properly, so
this is a speculative fix.

Bug: 447189642
Change-Id: I390998253e78ac110ef9015a74bc057a958a97cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7008799
Commit-Queue: Michael Wilson <mjwilson@chromium.org>
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1526443}