Fix TabStripGroupContextMenuTest flakiness on android-15-tablet-x64-rel
The flakiness came from the removed line that was used to arbitrarily tapped the screen to dismiss the menu. That sometimes tapped the sign-in promo card on the tablet bot, leaving the UI stuck on the sign-in prompt, so dismiss the menu directly on UI thread.

Screenshot(test stuck on sign-in prompt): https://screenshot.googleplex.com/An784HvSjqVxu7d

Bug: 447179264
Change-Id: Id2d3fbfbf1d638347bfece5acc0e83a28710e1f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7011636
Reviewed-by: Neil Coronado <nemco@google.com>
Commit-Queue: Neil Coronado <nemco@google.com>
Auto-Submit: Zhe Li <zheliooo@google.com>
Cr-Commit-Position: refs/heads/main@{#1526525}